app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.10